requirements.txt
# 웹 스크래핑
selenium==4.15.2
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
requests==2.31.0

# 데이터 처리
pandas==2.1.3
numpy==1.24.3
pillow==10.1.0

# 설정 관리
pyyaml==6.0.1
python-dotenv==1.0.0

# 로깅
loguru==0.7.2

# 웹 자동화 추가 도구
webdriver-manager==4.0.1

# 데이터베이스 (선택사항)
# sqlite3 - Python 표준 라이브러리에 포함

# 키보드 입력 처리 - 표준 input() 사용으로 제거
# keyboard==0.13.5

# 시간 관리
schedule==1.2.0

# 파일 처리
pathlib2==2.3.7.post1

# 이미지 처리
opencv-python==4.8.1.78

# 타입 힌트
typing-extensions==4.8.0